requirements.txt
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
python-pptx
langchain_community
google-generativeai
streamlit